Moved LazyOption from mod.rs into lazy_option.rs
